# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/health": {
      "get": {
        "operationId": "getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "API and Postgres liveness",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "postAuthLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid phone or PIN"
          },
          "403": {
            "description": "Account is suspended"
          },
          "429": {
            "description": "Too many PIN attempts or login rate limit"
          }
        },
        "summary": "Sign in with phone and PIN",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Accepts `refreshToken` in the JSON body or the `refresh_token` httpOnly cookie.",
        "operationId": "postAuthRefresh",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token missing or invalid"
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "Rotate the refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "postAuthLogout",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Access token missing or invalid"
          }
        },
        "security": [
          {
            "refresh_token": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Revoke the current refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/pin": {
      "post": {
        "operationId": "postAuthPin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePinDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Access token or current PIN invalid"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace PIN",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MeUserResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/MeStaffResponseDto"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "kind",
                    "mapping": {
                      "user": "#/components/schemas/MeUserResponseDto",
                      "staff": "#/components/schemas/MeStaffResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Access token missing or invalid"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Current user or staff session",
        "tags": [
          "me"
        ]
      }
    }
  },
  "info": {
    "title": "IoTeedom Smart Pay API",
    "description": "Auth, session, and health endpoints for Smart Pay.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "local"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "service": {
            "enum": [
              "api",
              "worker"
            ],
            "type": "string"
          },
          "postgres": {
            "enum": [
              "up",
              "down"
            ],
            "type": "string"
          },
          "redis": {
            "enum": [
              "up",
              "down"
            ],
            "type": "string"
          }
        },
        "required": [
          "ok",
          "service"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "024 412 8891",
            "description": "Ghana phone number."
          },
          "pin": {
            "type": "string",
            "pattern": "^\\d{4,6}$",
            "example": "2468",
            "description": "4 to 6 digit PIN."
          }
        },
        "required": [
          "phone",
          "pin"
        ]
      },
      "AuthTokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token. Send as `Authorization: Bearer <token>`."
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token. Also set as the httpOnly `refresh_token` cookie on `/v1/auth`."
          },
          "expiresIn": {
            "type": "number",
            "description": "Access token lifetime in seconds.",
            "example": 900
          },
          "mustChangePin": {
            "type": "boolean",
            "description": "True when the user must replace their temporary PIN."
          },
          "principal": {
            "enum": [
              "user",
              "staff"
            ],
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "mustChangePin",
          "principal"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token. Omit when the `refresh_token` cookie is present."
          }
        }
      },
      "ChangePinDto": {
        "type": "object",
        "properties": {
          "currentPin": {
            "type": "string",
            "pattern": "^\\d{4,6}$",
            "description": "Required unless `mustChangePin` is true (first login with a temp PIN)."
          },
          "newPin": {
            "type": "string",
            "pattern": "^\\d{4,6}$",
            "description": "4 to 6 digit PIN."
          }
        },
        "required": [
          "newPin"
        ]
      },
      "MeMembershipDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "owner",
              "manager"
            ],
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "accountKind": {
            "enum": [
              "home",
              "estate"
            ],
            "type": "string"
          },
          "modules": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            },
            "example": {
              "ecg": true,
              "water": true,
              "utilities": false,
              "meters": false,
              "smart_home": false,
              "solar": false,
              "ev": false
            }
          }
        },
        "required": [
          "role",
          "accountId",
          "accountName",
          "accountKind",
          "modules"
        ]
      },
      "MeUserResponseDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "user"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "233244128891"
          },
          "mustChangePin": {
            "type": "boolean"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeMembershipDto"
            }
          }
        },
        "required": [
          "kind",
          "id",
          "name",
          "phone",
          "mustChangePin",
          "memberships"
        ]
      },
      "MeStaffResponseDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "staff"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "233200000001"
          },
          "role": {
            "enum": [
              "superadmin",
              "support"
            ],
            "type": "string"
          },
          "mustChangePin": {
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "id",
          "name",
          "phone",
          "role",
          "mustChangePin"
        ]
      }
    }
  }
}